packages/core/requirements.txt
# BUDDY Core Runtime Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
asyncio-throttle>=1.0.0

# Event processing
redis>=5.0.0  # Optional for distributed event bus
nats-py>=2.6.0  # Alternative event bus

# Configuration and data
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Database
motor>=3.3.0
pymongo>=4.6.0
dnspython>=2.4.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional ML dependencies (for local models)
torch>=2.1.0; sys_platform != "win32"
transformers>=4.35.0
sentence-transformers>=2.2.0

# Audio processing (will be used by voice package)
numpy>=1.24.0
soundfile>=0.12.0
